Add unit test for more requests and responses
